# backend/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@tesla-app.com
    name: API Support
    url: http://localhost:8080
  description: API for Tesla driving scenarios and trip management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Tesla App API
  version: "1.0"
paths:
...
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
      summary: User login
      tags:
      - Users
  /api/users/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate token
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Users
  /api/users/profile:
    get:
      description: Get current user profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update current user profile information
      parameters:
      - description: Profile update data
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User registration data
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
      summary: User registration
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"